change datafetching in App and pass data to dashboardStore, using dashbordStore data in dashboard.vue component
